Etap 3: izolacja i odporność — wyjątki z Ring 3, FPU per zadanie, W^X, strony ochronne, nowa sterta, autotest

+++ kernel/ui/ui_scale.c
@@ -1,4 +1,4 @@
 #include "ui_scale.h"
 
-/* Set by kernel_main from the video mode: 2.0 on 1920x1080, else 1.0 */
-float ui_scale = 2.0f;
+/* Set by kernel_main from the video mode: 2 on 1920x1080, else 1 */
+int ui_scale = 2;
